requirements.txt
# Core MCP dependencies
mcp
python-dotenv
aiohttp

# Google APIs (GSC, YouTube)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0

# Social platforms
mastodon.py>=1.5.0
atproto>=0.0.0
requests>=2.28.0
requests-oauthlib>=1.3.0
facebook-sdk>=3.0.0

# Data handling
pydantic>=2.0.0
sqlite-utils>=3.0.0
